+++ CHANGELOG.md
@@ -13,6 +13,8 @@ Given a version number MAJOR.MINOR.PATCH, increment:
 
 
 ## [Unreleased]
+### Added
+- reversalDisplayDescription to Deposit resource
 
 ## [2.33.0] - 2025-07-08
 ### Added